rename: Atom 네이밍 변경
chat_atom, room_atom 같이 명확하지 않은 이름에서
더 명확한 이름으로 명시함.
