+++ src/pages/interviewRoom.jsx
@@ -1,7 +1,7 @@
 import {useNavigate} from "react-router-dom";
 import {useEffect} from "react";
 import {useRecoilState} from "recoil";
-import {roomIdAtom} from "../store/room_atom";
+import {roomIdAtom} from "../store/interviewRoomAtom";
 import InterviewInput from "./interviewRoom/interviewInput";
 import InterviewChat from "./interviewRoom/interviewChat";
 import InterviewFeedback from "./interviewRoom/interviewFeedback";
