+++ src/pages/interviewRoom/interviewFeedback.jsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import style from "../../styles/interviewFeedback.module.css";
 import {useRecoilState} from "recoil";
-import {interviewDataAtom} from "../../store/room_atom";
+import {interviewDataAtom} from "../../store/interviewRoomAtom";
 import "chart.js/auto";
 import { Radar } from "react-chartjs-2";
 import {FEEDBACK_RANGE_DEFAULT_VALUE} from "../../constants/interviewFeedbackConst";
